+++ applications/Unity.GrantManager/modules/Unity.Identity.Web/src/Pages/Identity/Users/ImportModal.cshtml.cs
@@ -3,10 +3,8 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Volo.Abp.ObjectExtending;
 using Volo.Abp.Identity;
-using Unity.GrantManager.UserImport;
+using Unity.GrantManager.Identity;
 using System.ComponentModel;
-using Microsoft.Extensions.Localization;
-using Unity.GrantManager.Localization;
 
 namespace Unity.Identity.Web.Pages.Identity.Users;
