+++ applications/Unity.GrantManager/src/Unity.GrantManager.Application.Contracts/Identity/UserSearchDto.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Unity.GrantManager.UserImport
+namespace Unity.GrantManager.Identity
 {
     public class UserSearchDto
     {
