+++ applications/Unity.GrantManager/src/Unity.GrantManager.DbMigrator/appsettings.json
@@ -1,7 +1,7 @@
 {
   "ConnectionStrings": {
-    "Default": "Host=localhost;port=5433;Database=UnityGrantManager;Integrated Security=True;Username=postgres;Password=admin",
-    "Tenant": "Host=localhost;port=5433;Database=UnityGrantTenant;Integrated Security=True;Username=postgres;Password=admin"
+    "Default": "Host=localhost;port=5432;Database=UnityGrantManager;Integrated Security=True;Username=postgres;Password=admin",
+    "Tenant": "Host=localhost;port=5432;Database=UnityGrantTenant;Integrated Security=True;Username=postgres;Password=admin"
   },
   "Redis": {
     "Configuration": "127.0.0.1"
