+++ applications/Unity.GrantManager/test/Unity.GrantManager.Application.Tests/Identity/UserImportAppServiceTests.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Unity.GrantManager.Identity
+{
+    internal class UserImportAppServiceTests
+    {
+    }
+}
